% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodfuns.R
\name{coef.pcoxtime}
\alias{coef.pcoxtime}
\alias{coefficients.pcoxtime}
\title{Extract coefficient estimates of pcoxtime object}
\usage{
\method{coef}{pcoxtime}(object, ...)

\method{coefficients}{pcoxtime}(object, ...)
}
\arguments{
\item{object}{fitted \code{\link[pcoxtime]{pcoxtime}} model object}

\item{...}{for future implementations}
}
\value{
A vector of coefficient estimates.

A vector of coefficient estimates.
}
\description{
This function extracts the estimates for all the coefficients.
}
\details{
The call that produced \code{\link[pcoxtime]{pcoxtime}} is printed, followed by coefficient estimates.
}
