% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthocfuns.R
\name{extractoptimal.pcoxtimecv}
\alias{extractoptimal.pcoxtimecv}
\alias{extractoptimal}
\title{Extract optimal parameter values}
\usage{
\method{extractoptimal}{pcoxtimecv}(object, what = c("optimal", "cvm", "coefs"), ...)
}
\arguments{
\item{object}{\code{\link[pcoxtime]{pcoxtimecv}} object}

\item{what}{the summary or data frame to extract. Currently, three options are available: \code{what = "optimal"} extracts a data frame showing optimal parameter values, \code{what = "cvm"} extracts the data frame containing the mean cross-validation error for various lambda-alpha combination, and \code{what = "coef"}, requires \code{refit = TRUE} in \code{\link[pcoxtime]{pcoxtimecv}},  extracts a data frame containing the coefficient estimates for various lambda-alpha combination.}

\item{...}{for future implementations}
}
\value{
A data frame depending on the specification described above.
}
\description{
Extract cross-validation summaries and data frames.
}
\details{
Extract cross-validation summaries based on the optimal parameters or data frames containing all the summaries for all the parameter values.
}
