% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfuns.R
\name{getResponse}
\alias{getResponse}
\title{Extract response from the formula}
\usage{
getResponse(formula)
}
\arguments{
\item{model}{formula.}
}
\description{
Extract response from the formula
}
\keyword{internal}
