% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfuns.R
\name{parseFormula}
\alias{parseFormula}
\title{Parse formula and return response variable.}
\usage{
parseFormula(formula, data, env = parent.frame())
}
\arguments{
\item{formula}{Object of class formula describing 
the model.}

\item{data}{optional data frame containing.
variables specified in the formula.}
}
\value{
A character string. The name of the response variable
}
\description{
Parse formula and return response variable.
}
\keyword{internal}
