% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoxtheme.R
\name{pcoxtheme}
\alias{pcoxtheme}
\title{Set theme for pcoxtime plots}
\usage{
pcoxtheme()
}
\value{
No return value, called for side effects (setting pcotime plotting theme).
}
\description{
Sets a theme for pcoxtime and other ggplot objects
}
\examples{
library(ggplot2)
pcoxtheme()
data(heart, package="survival")
lam <- 0.02
alp <- 1
pfit <- pcoxtime(Surv(start, stop, event) ~ age + year + surgery + transplant
	, data = heart
	, lambda = lam
	, alpha = alp
)

# Plot survival curves
psurv <- pcoxsurvfit(pfit)
plot(psurv)
}
