% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoxtimeplots.R
\name{plot.varimp}
\alias{plot.varimp}
\title{Generic method for plotting variable importance}
\usage{
\method{plot}{varimp}(x, ..., pos = 0.5, drop_zero = TRUE)
}
\arguments{
\item{x}{a \code{\link[pcoxtime]{varimp}} object.}

\item{...}{for future implementations.}

\item{pos}{spacing between labels.}

\item{drop_zero}{if \code{TRUE} only nonzero estimates are shown.}
}
\description{
Plots variable importance for \code{\link[pcoxtime]{pcoxtime}} fit.
}
\seealso{
\code{\link[pcoxtime]{varimp}}
}
