% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthocfuns.R
\name{predictRisk.pcoxtime}
\alias{predictRisk.pcoxtime}
\alias{predictRisk}
\title{Extract predictions from pcoxtime model}
\usage{
\method{predictRisk}{pcoxtime}(object, newdata, times, ...)
}
\arguments{
\item{object}{fitted \code{\link[pcoxtime]{pcoxtime}}.}

\item{newdata}{a data frame containing the variables appearing on the right hand side of \code{\link[pcoxtime]{pcoxtime}} formula.}

\item{times}{a vector of times in the range of the response, at which to return the survival probabilities.}

\item{...}{for future implementations.}
}
\value{
a matrix of probabilities with as many rows as the rows of the \code{newdata} and as many columns as number of time points (\code{times}).
}
\description{
Extract event probabilities from the fitted model.
}
\details{
For survival outcome, the function predicts the risk, \eqn{1 - S(t|x)}, where \eqn{S(t|x)} is the survival chance of an individual characterized by \eqn{x}.
}
\examples{

data(veteran, package="survival")
# Penalized
lam <- 0.1
alp <- 0.5
pfit1 <- pcoxtime(Surv(time, status) ~ factor(trt) + karno + diagtime + age + prior
	, data = veteran
	, lambda = lam 
	, alpha = alp
)
r1 <- predictRisk(pfit1, newdata = veteran[1:80,], times = 10)

# Unpenalized
lam <- 0
alp <- 1
pfit2 <- pcoxtime(Surv(time, status) ~ factor(trt) + karno + diagtime + age + prior
	, data = veteran
	, lambda = lam 
	, alpha = alp
)
r2 <- predictRisk(pfit2, newdata = veteran[1:80,], times = 10)
plot(r1, r2, xlim=c(0,1), ylim=c(0,1)
	, xlab = "Penalized predicted survival chance at 10"
	, ylab="Unpenalized predicted survival chance at 10"
)

}
