% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodfuns.R
\name{print.pcoxbasehaz}
\alias{print.pcoxbasehaz}
\title{Print baseline hazard function data frame}
\usage{
\method{print}{pcoxbasehaz}(x, n = 5, ...)
}
\arguments{
\item{x}{the result of a call to the \code{\link[pcoxtime]{pcoxbasehaz.pcoxtime}} function.}

\item{n}{number of rows to print. Default is 5.}

\item{...}{for future implementations}
}
\value{
The call to the \code{\link[pcoxtime]{pcoxbasehaz.pcoxtime}} and the head of baseline hazard function data frame.
}
\description{
Print the head of baseline hazard function data frame.
}
\details{
Provide a summary of \code{\link[pcoxtime]{pcoxbasehaz.pcoxtime}} object.
}
