% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodfuns.R
\name{print.pcoxsurvfit}
\alias{print.pcoxsurvfit}
\title{Print a short summary of survival function}
\usage{
\method{print}{pcoxsurvfit}(x, ...)
}
\arguments{
\item{x}{the result of a call to the \code{\link[pcoxtime]{pcoxsurvfit.pcoxtime}} function.}

\item{...}{for future implementations}
}
\value{
The call to the \code{\link[pcoxtime]{pcoxsurvfit.pcoxtime}} and the summary of the survival function.
}
\description{
Print the number of observations and number of events.
}
\details{
Provide a summary of \code{\link[pcoxtime]{pcoxsurvfit.pcoxtime}} object.
}
