% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodfuns.R
\name{print.pcoxtimecv}
\alias{print.pcoxtimecv}
\title{Print cross-validated pcoxtime object}
\usage{
\method{print}{pcoxtimecv}(x, ...)
}
\arguments{
\item{x}{\code{\link[pcoxtime]{pcoxtimecv}} object}

\item{...}{for future implementations}
}
\value{
The call to the \code{\link[pcoxtime]{pcoxtimecv}} and the summary of the optimal alpha and lambdas.
}
\description{
Print the summary of the result of cross-validation for a pcoxtime object.
}
\details{
A summary of optimal lambda and alpha for training pcoxtime model.
}
